Insert, Edit, Delete Terms

+++ client/src/pages/ManagementTerm/ManagementTerm.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Button, Dropdown, Modal, Form } from 'react-bootstrap';
+import { Button, Dropdown, Modal, Form, Col } from 'react-bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import "../ManagementTerm/ManagementTerm.css";
 import { useNavigate } from "react-router-dom";
@@ -8,19 +8,22 @@ type Term = {
     terms_id: string;
     terms_title: string;
     terms_content: string;
+    terms_mandatory: boolean;
 }
 
 const ManagementTerm: React.FC = () => {
     const [terms, setTerms] = useState<Term[]>([]);
     const [showModal, setShowModal] = useState(false);
+    const [showConfirmModal, setShowConfirmModal] = useState(false); // Novo estado
     const [termTitle, setTermTitle] = useState('');
     const [termBody, setTermBody] = useState('');
+    const [termIsMandatory, setTermIsMandatory] = useState<Boolean | null>(null); // Alterado para permitir null
     const [isUpdate, setIsUpdate] = useState(false);
     const [currentTerm, setCurrentTerm] = useState<Term | null>(null);
 
     const fetchTerms = async () => {
         try {
-            const response = await fetch(`http://localhost:3001/Terms/GetTerms`, {
+            const response = await fetch('http://localhost:3001/Terms/GetTerms', {
                 method: 'GET'
             });
             if (!response.ok) {
@@ -50,58 +53,88 @@ const ManagementTerm: React.FC = () => {
 
     const handleAddTerm = () => {
         setIsUpdate(false);
+        setTermTitle('');
+        setTermBody('');
+        setTermIsMandatory(null); // Define como null ao adicionar um novo termo
         setShowModal(true);
     };
 
     const handleAttTerm = (term: Term) => {
         setCurrentTerm(term);
+        setTermTitle(term.terms_title);
+        setTermBody(term.terms_content);
+        setTermIsMandatory(term.terms_mandatory);
         setIsUpdate(true);
         setShowModal(true);
     };
 
     const handleUpload = async () => {
-        if (!termTitle || !termBody) {
+        if (!termTitle || !termBody || termIsMandatory === null) { // Verifica se o termIsMandatory não é null
             alert('Por favor, preencha todos os campos.');
             return;
         }
-    
+
+        const url = isUpdate ? `http://localhost:3001/Terms/PutTerms/${currentTerm?.terms_id}` : 'http://localhost:3001/Terms/AddTerm';
+        const method = isUpdate ? 'PUT' : 'POST';
+
         try {
-            const response = await fetch('http://localhost:3001/Terms/AddTerm', {
-                method: 'POST',
+            const response = await fetch(url, {
+                method: method,
                 headers: {
                     'Content-Type': 'application/json'
                 },
-                body: JSON.stringify({ terms_title: termTitle, terms_content: termBody })
+                body: JSON.stringify({ terms_title: termTitle, terms_content: termBody, terms_mandatory: termIsMandatory })
             });
-    
+
             if (!response.ok) {
-                const errorData = await response.json();
-                throw new Error(errorData.message); // Use a propriedade correta do objeto de erro retornado pelo backend
+                const errorText = await response.text();
+                throw new Error('Erro ao adicionar/atualizar termo: ' + errorText);
             }
-    
-            setShowModal(false);
-            fetchTerms(); // Atualiza a lista de termos
-        } catch (error: any) {
+
+            const responseData = await response.json();
+            if (responseData.success) {
+                setShowModal(false);
+                await fetchTerms();
+            } else {
+                throw new Error('Erro ao adicionar/atualizar termo: ' + responseData.message);
+            }
+        } catch (error: unknown) {
             console.error(error);
-            alert('Erro ao adicionar/atualizar termo: ' + error.message);
+            if (error instanceof Error) {
+                alert(error.message);
+            } else {
+                alert('Ocorreu um erro desconhecido.');
+            }
         }
     };
 
-    const handleDeleteTerm = async (termId: string) => {
+    const handleDeleteTerms = (termId: string) => {
+        setCurrentTerm(terms.find(term => term.terms_id === termId) || null);
+        setShowConfirmModal(true);
+    };
+
+    const handleConfirmDeleteTerm = async () => {
+        if (!currentTerm) return;
+
         try {
-            const response = await fetch(`http://localhost:3001/Terms/DeleteTerm/${termId}`, {
+            const response = await fetch(`http://localhost:3001/DeleteTerms/DeleteTerms/${currentTerm.terms_id}`, {
                 method: 'DELETE'
             });
 
             if (!response.ok) {
-                
-                throw new Error('Erro ao excluir termo');
+                const errorText = await response.text();
+                throw new Error('Erro ao excluir termo: ' + errorText);
             }
 
-            fetchTerms(); // Atualiza a lista de termos após a exclusão
-        } catch (error) {
+            setShowConfirmModal(false);
+            await fetchTerms();
+        } catch (error: unknown) {
             console.error(error);
-            alert('Erro ao excluir termo');
+            if (error instanceof Error) {
+                alert(error.message);
+            } else {
+                alert('Erro ao excluir termo');
+            }
         }
     };
 
@@ -125,17 +158,17 @@ const ManagementTerm: React.FC = () => {
                     <div className="table-container">
                         <table className="wide-table">
                             <tbody>
-                                {terms && terms.map && terms.map((term, index) => (
+                                {terms.map((term, index) => (
                                     <tr key={index}>
                                         <th>{term.terms_title}</th>
                                         <td>
                                             <Dropdown>
-                                                <Dropdown.Toggle variant="primary" id="dropdown-basic">
+                                                <Dropdown.Toggle variant="primary" id={`dropdown-basic-${index}`}>
                                                     Ações
                                                 </Dropdown.Toggle>
                                                 <Dropdown.Menu>
                                                     <Dropdown.Item onClick={() => handleAttTerm(term)}>Atualizar Termo</Dropdown.Item>
-                                                    <Dropdown.Item className="logout-item" onClick={() => handleDeleteTerm(term.terms_id)}>Excluir Termo</Dropdown.Item>
+                                                    <Dropdown.Item className="logout-item" onClick={() => handleDeleteTerms(term.terms_id)}>Excluir Termo</Dropdown.Item>
                                                 </Dropdown.Menu>
                                             </Dropdown>
                                         </td>
@@ -161,6 +194,27 @@ const ManagementTerm: React.FC = () => {
                             <Form.Label>Corpo do Termo de Uso</Form.Label>
                             <Form.Control as="textarea" rows={5} value={termBody} onChange={(e) => setTermBody(e.target.value)} />
                         </Form.Group>
+                        <Form.Group controlId="formTermIsMandatory">
+                            <Form.Label>O termo é obrigatório? </Form.Label>
+                            <Col>
+                                <Form.Check
+                                    type="radio"
+                                    label="Sim"
+                                    name="termIsMandatory"
+                                    value={'true'}
+                                    checked={termIsMandatory === true}
+                                    onChange={() => setTermIsMandatory(true)}
+                                />
+                                <Form.Check
+                                    type="radio"
+                                    label="Não"
+                                    name="termIsMandatory"
+                                    value={'false'}
+                                    checked={termIsMandatory === false}
+                                    onChange={() => setTermIsMandatory(false)}
+                                />
+                            </Col>
+                        </Form.Group>
                     </Form>
                 </Modal.Body>
                 <Modal.Footer>
@@ -172,8 +226,26 @@ const ManagementTerm: React.FC = () => {
                     </Button>
                 </Modal.Footer>
             </Modal>
+            {/* Modal de confirmação para excluir termo */}
+            <Modal show={showConfirmModal} onHide={() => setShowConfirmModal(false)}>
+                <Modal.Header closeButton>
+                    <Modal.Title>Confirmar Exclusão</Modal.Title>
+                </Modal.Header>
+                <Modal.Body>
+                    Tem certeza de que deseja excluir o termo?
+                </Modal.Body>
+                <Modal.Footer>
+                    <Button variant="secondary" onClick={() => setShowConfirmModal(false)}>
+                        Não
+                    </Button>
+                    <Button variant="primary" onClick={handleConfirmDeleteTerm}>
+                        Sim
+                    </Button>
+                </Modal.Footer>
+            </Modal>
         </div>
     );
 }
 
 export default ManagementTerm;
+
